// src/models/Enum.ts
export const faceValues = ["A", "2", "3", "4", "5", "6", "7", "8", "9", "T", "J", "Q", "K"];

export enum Suit {
    H,
    D,
    S,
    C,
}

export enum HandType {
    High_Card = "High card",
    One_Pair = "One pair",
    Two_Pair = "Two pair",
    Three_Of_A_Kind = "Three of a kind",
    Straight = "Straight",
    Flush = "Flush",
    Full_House = "Full house",
    Four_Of_A_Kind = "Four of a kind",
    Straight_Flush = "Straight flush",
    Royal_Flush = "Royal Flush"
}